Refactor svg icons and move them to stateless components for modal-ui.
