+++ packages/modal-ui/src/lib/components/CloseButton.tsx
@@ -1,4 +1,5 @@
 import React from "react";
+import { CloseIcon } from "./icons/CloseIcon";
 
 interface CloseButtonProps {
   onClick: () => void;
@@ -7,16 +8,7 @@ interface CloseButtonProps {
 export const CloseButton: React.FC<CloseButtonProps> = ({ onClick }) => {
   return (
     <button onClick={onClick} className="close-button">
-      <svg
-        xmlns="http://www.w3.org/2000/svg"
-        height="24"
-        viewBox="0 0 24 24"
-        width="24"
-        fill="#C1C1C1"
-      >
-        <path d="M0 0h24v24H0z" fill="none" />
-        <path d="M19 6.41L17.59 5 12 10.59 6.41 5 5 6.41 10.59 12 5 17.59 6.41 19 12 13.41 17.59 19 19 17.59 13.41 12z" />
-      </svg>
+      <CloseIcon />
     </button>
   );
 };
