+++ packages/modal-ui/src/lib/components/icons/BackArrowIcon.tsx
@@ -1,0 +1,21 @@
+import React from "react";
+
+export const BackArrowIcon: React.FC = () => {
+  return (
+    <svg
+      width="8"
+      height="14"
+      viewBox="0 0 8 14"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M7 13L1 7L7 1"
+        stroke="#6494EE"
+        strokeWidth="1.5"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+      />
+    </svg>
+  );
+};
