+++ packages/modal-ui/src/lib/components/icons/CopyIcon.tsx
