+++ packages/modal-ui/src/lib/components/icons/DownArrowIcon.tsx
@@ -1,0 +1,21 @@
+import React from "react";
+
+export const DownArrowIcon: React.FC = () => {
+  return (
+    <svg
+      width="10"
+      height="7"
+      viewBox="0 0 10 7"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M1 1.52382L5 5.52382L9 1.52382"
+        stroke="#4F7CD1"
+        strokeWidth="1.5"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+      />
+    </svg>
+  );
+};
