+++ packages/modal-ui/src/lib/components/icons/FolderIcon.tsx
@@ -1,0 +1,27 @@
+import React from "react";
+
+export const FolderIcon = () => (
+  <svg
+    width="40"
+    height="41"
+    viewBox="0 0 40 41"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <circle cx="28.3333" cy="23.8333" r="1.66667" fill="currentColor"></circle>
+    <path
+      d="M35 12.1667H7C5.89543 12.1667 5 11.2712 5 10.1667V7.5C5 6.39543 5.89543 5.5 7 5.5H31.6667"
+      stroke="currentColor"
+      strokeWidth="3"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+    ></path>
+    <path
+      d="M35 12.1667V35.5H7C5.89543 35.5 5 34.6046 5 33.5V8.83334"
+      stroke="currentColor"
+      strokeWidth="3"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+    ></path>
+  </svg>
+);
