+++ packages/modal-ui/src/lib/components/icons/LedgerDeviceIcon.tsx
