+++ packages/modal-ui/src/lib/components/icons/QRIcon.tsx
@@ -1,0 +1,57 @@
+import React from "react";
+
+export const QRIcon = () => (
+  <svg
+    width="18"
+    height="16"
+    viewBox="0 0 18 16"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path
+      d="M7.22224 1.33334H1.44446V6.66668H7.22224V1.33334Z"
+      stroke="currentColor"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+    />
+    <path
+      d="M15.8889 1.33334H10.1111V6.66668H15.8889V1.33334Z"
+      stroke="currentColor"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+    />
+    <path
+      d="M7.22224 9.33334H1.44446V14.6667H7.22224V9.33334Z"
+      stroke="currentColor"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+    />
+    <path
+      d="M10.1111 13.1429V14.6667H15.8889M10.1111 9.33334V10.8572H12.5873V9.33334H15.8889V12.381"
+      stroke="#4C5155"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+    />
+    <rect
+      x="3.61108"
+      y="3.33334"
+      width="1.44444"
+      height="1.33333"
+      fill="#4C5155"
+    />
+    <rect
+      x="3.61108"
+      y="11.3333"
+      width="1.44444"
+      height="1.33333"
+      fill="#4C5155"
+    />
+    <rect
+      x="12.2778"
+      y="3.33334"
+      width="1.44445"
+      height="1.33333"
+      fill="#4C5155"
+    />
+  </svg>
+);
