+++ packages/modal-ui/src/lib/components/icons/UpArrowIcon.tsx
@@ -1,0 +1,21 @@
+import React from "react";
+
+export const UpArrowIcon: React.FC = () => {
+  return (
+    <svg
+      width="10"
+      height="7"
+      viewBox="0 0 10 7"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M9 5.4762L5 1.4762L1 5.4762"
+        stroke="#4F7CD1"
+        strokeWidth="1.5"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+      />
+    </svg>
+  );
+};
